requirements.txt
jsonlines
pandas>=2.0.0
numpy>=1.24.0
tabulate>=0.9.0
typing-extensions>=4.5.0
openai>=1.0.0
dspy
dspy[optuna]
duckdb
python-dotenv
vllm==0.8.3
scikit-learn
PyYAML